requirements.txt
# Credit Intelligence - Requirements
# Python 3.10+ recommended

# =============================================================================
# LangChain & LangGraph (Core)
# =============================================================================
langgraph>=0.2.0,<1.0.0
langchain>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-openai>=0.2.0,<1.0.0
langchain-anthropic>=0.2.0,<1.0.0
langsmith>=0.1.0

# =============================================================================
# LLM Providers
# =============================================================================
groq>=0.9.0
openai>=1.0.0
anthropic>=0.25.0

# =============================================================================
# Database & Storage
# =============================================================================
pymongo>=4.6.0
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# =============================================================================
# HTTP & APIs
# =============================================================================
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.0

# =============================================================================
# Data Processing
# =============================================================================
pydantic>=2.5.0,<3.0.0
pyyaml>=6.0.0

# =============================================================================
# Web Search
# =============================================================================
duckduckgo-search>=6.0.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
rich>=13.0.0
tqdm>=4.66.0

# =============================================================================
# Optional: For web frontend (Vercel deployment)
# =============================================================================
# fastapi>=0.109.0
# uvicorn>=0.27.0
# sse-starlette>=2.0.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
